+++ CHANGELOG.rst
@@ -10,9 +10,15 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 **Added**
 
+Introduce life.qbic.utils.BioinformaticAnalysisParser class to parse and validate the filestructure resulting from Nfcore pipeline output (`#51 <https://github.com/qbicsoftware/core-utils-lib/pull/51>`_)
+
 **Fixed**
 
 **Dependencies**
+
+Update data-model-lib to 2.7.0 (`#51 <https://github.com/qbicsoftware/core-utils-lib/pull/51>`_)
+
+Re-introduce org.apache.commons.commons-lang3 dependency with version 3.7.0
 
 **Deprecated**